Extract overhead SVG plot.
